refactor: rename cfd-compare-accounts to compare-accounts
