+++ packages/cfd/src/Containers/cfd-compare-accounts/compare-accounts.scss
