+++ packages/cfd/src/Containers/cfd-compare-accounts/index.tsx
@@ -1,4 +1,4 @@
-import CFDCompareAccounts from './cfd-compare-accounts';
-import './cfd-compare-accounts.scss';
+import CFDCompareAccounts from './compare-accounts';
+import './compare-accounts.scss';
 
 export default CFDCompareAccounts;
